Remove impl from user defined entity names
