+++ dds/src/implementation/dds_impl/user_defined_data_reader_node.rs
@@ -1,0 +1,2 @@
+#[derive(PartialEq, Debug)]
+pub struct UserDefinedDataReaderNode;
